multiple-pulse capability enabled
The capability to emit multiple pulses per sequence step was added to the driver. For this, the sweep mode was misused by setting start and stop voltage/current to the same value. 

As this also works for a single pulse as well as the DC mode, the initialization was modified to have the B290xB SMU run in the sweep mode all the time, even if the pulse count in the GUI is set to 1 or if the SMU is used for DC output. No drawbacks were observed in practical use other than that the symbol on the display of the SMU shows a sweep rather than a DC or pulse graphic.