main.py for Keysight EXR/MXR/UXR Oscilloscopes
This driver was tested on a Keysight EXR scope, but MXR and UXR should work fine as well as only basic functionality was implemented and all three scopes run on the same software.

Development of the driver began with the Tek7k driver as a starting point but just few pieces are left. Beside configuration of channels, trigger and sample rate, the sample rate type option was misused to implement the possibility of ERES, offering the tradin of bandwidth and sample rate for an increase in ENOB, up to 16bit.

Some wait times had to be inserted to take care of processing time on the scope. Especially if you start the scope with default settings and request it to change pretty much everything from vertical scale to horizontal scale, trigger type, trigger edge, etc. etc. it could be that it triggers before finalizing on all changes, throwing an "invalid waveform" error because the captured one in memory does not fit the requested parameters.

The data transfer uses ASCii as it is nice to handle during debugging and requires not processing after retrival. If speed is a concern, a switch to WORD format including processing of the values needs to be implemented.