initial driver release for HP/Agilent/Keysight E3631A Triple-Output PSU
Based on the E3632A driver but some changes were necessary to adjust to the E3631A command set.
The so-called "TRACK" mode is available as well as a channel selection, synchronizing the +25V and -25V to a common absolute voltage value for e.g. OPA supply. In TRACK mode, the driver is programmed to accept positive voltage values to output this on the +25V terminal and the respective negative equivalent to the -25V terminal.
As usual, some sanity checks including error messages are implemented as well regarding maximum voltage and current limits and polarity of input.