adding measurement in DC mode also into 'measure' section
After deleting the exidentially created 'patch-1' branch in my repo I would like to continue the discussion about the move of the pulse emission into the 'measure' section of the driver.

IMHO, the move of the pulse emission is justified as the measurement is taken simultaniously and therefore belongs foremost to the 'measure' section instead of the 'apply' section. It is therefore a good idea.

However, we should definitly keep the structure conclusive by moving the measurement of the supplied current and voltage in DC mode into the 'measure' section of the driver as well while leaving the application of the output in the 'apply' section as measurement and application are seperated in time for DC, contrary to the Pulse mode.

This commit takes care of the move of the measurement in DC mode.