HP 3456A DMM/Logger driver
Driver for the HP 3456A DMM, essentially a 6.5 Volt- and Ohmmeter, no current measurement capabilities.
Besides DCV, ACV, DC+ACV, 2-Wire and 4-Wire resistance, the driver also includes the (O)ffset (C)ompensated measurement mode for 2W and 4W Resistance measurements.
Number of NPLC, Auto-Zero and analogue Filter options are available, too.
The display option does not actually switch of the display but just switches every digit to display an uppercase "H" to hide the values. No idea if this is of use, but it is implemented as a native instrument function anyway.

Note a few specialities: 
- this is pre-SCPI, the HP-IB commands are few and short. 
- there is no command to request a measurement, rather the current result can be retrieved from the instrument just by reading the bus
- the device has a hardware-enabled talking-only mode to feed a printer or similar; watch out that it is not set as it won't accept commands via GPIB while in this mode, which is set via the GBPI address dip switches (see driver comment)

The 3456A is known for its superb long-time stability and used even today as a lab-internal reference in many places.