Single-Triggered Measurement added + some default settings
Trigger is put on hold now during 'configure' and set back to internal trigger during 'unconfigure'. A single trigger is released during measurement to have the point in time properly synced within the SweepMe sequencer.

The original driver assumed some parameters to be in standard configuration, the new once writes them within 'configure' to just make sure they are set as expected. The parameters are the End-Or-Identify mode and the output data format.

During the extension of the driver, I noticed the possibility for a mishap which I need to explain: on the frontpanel, one usually uses EITHER the specific MODE buttons ONLY to choose some of the functions modes like e.g. "DCV" OR one presses the SHIFT button first and then a certain function mode button to select the second function mode that is available only when SHIFT was pressed before, like e.g. 'O.C. 2W Resistance'.

Now, if you would press the SHIFT button on the device first and then start the program and select an "unSHIFTED" function mode, it would still choose the shifted function because the SHIFT button press is still in memory.

This new revision driver explicitly DE-selects SHIFT for the function modes in question by first writing 'S0' prior to sending the mode selection command.