Initial driver release HP 3457A DMM
Initial driver release for the HP 3457A DMM, successor to the previously released 3456A. The 3457A added the support for current measurement. In addition, it has a trick uppon his sleeve: despite being a 6.5 digit DMM, a 7th digit can be retrieved via GPIB. This comes in handy when making averaged readings on a constant value where the native noise gets reduced and the additional digit can be of real use, therefore the option was implemented.